+++ src/main/java/net/alcaris/plugin/items/enums/ArmorStatData.java
@@ -7,18 +7,18 @@ import java.util.Map;
  * 防具ステータスデータを管理するヘルパークラス
  */
 public class ArmorStatData {
-    private final Map<ArmorStats, Double> finalValues = new HashMap<>();
+    private final Map<ArmorStats, Integer> finalValues = new HashMap<>();
     private final Map<ArmorStats, Integer> performanceValues = new HashMap<>();
 
     public ArmorStatData() {
         // 初期化
         for (ArmorStats stat : ArmorStats.values()) {
-            finalValues.put(stat, 0.0);
+            finalValues.put(stat, 0);
             performanceValues.put(stat, 0);
         }
     }
 
-    public void setFinalValue(ArmorStats stat, double value) {
+    public void setFinalValue(ArmorStats stat, int value) {
         finalValues.put(stat, value);
     }
 
@@ -26,15 +26,15 @@ public class ArmorStatData {
         performanceValues.put(stat, value);
     }
 
-    public double getFinalValue(ArmorStats stat) {
-        return finalValues.getOrDefault(stat, 0.0);
+    public int getFinalValue(ArmorStats stat) {
+        return finalValues.getOrDefault(stat, 0);
     }
 
     public int getPerformanceValue(ArmorStats stat) {
         return performanceValues.getOrDefault(stat, 0);
     }
 
-    public Map<ArmorStats, Double> getFinalValues() {
+    public Map<ArmorStats, Integer> getFinalValues() {
         return new HashMap<>(finalValues);
     }
 
@@ -45,7 +45,7 @@ public class ArmorStatData {
     /**
      * 配列から値を設定
      */
-    public void setFromArrays(double[] finalValues, int[] performanceValues) {
+    public void setFromArrays(int[] finalValues, int[] performanceValues) {
         ArmorStats[] stats = ArmorStats.values();
         for (int i = 0; i < stats.length && i < finalValues.length && i < performanceValues.length; i++) {
             this.finalValues.put(stats[i], finalValues[i]);
@@ -56,11 +56,11 @@ public class ArmorStatData {
     /**
      * 配列として取得
      */
-    public double[] getFinalValuesAsArray() {
+    public int[] getFinalValuesAsArray() {
         ArmorStats[] stats = ArmorStats.values();
-        double[] result = new double[stats.length];
+        int[] result = new int[stats.length];
         for (int i = 0; i < stats.length; i++) {
-            result[i] = finalValues.getOrDefault(stats[i], 0.0);
+            result[i] = finalValues.getOrDefault(stats[i], 0);
         }
         return result;
     }
